# docs/openapi.yaml
openapi: 3.0.3
info:
  title: My Interview API
  version: 1.0.0
  description: Frontend contract verified against the current Spring Boot implementation.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/interview-sessions/{sessionId}:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    get:
      tags: [Interview]
      operationId: getInterviewSession
      summary: Get interview session status
      responses:
        '200':
          description: Session state.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewSessionStatus'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/interview-sessions/{sessionId}/preparation/retry:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    post:
      tags: [Interview]
      operationId: retryInterviewPreparation
      summary: Retry failed interview preparation
      responses:
        '202':
          description: Preparation redispatched.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewSessionStatus'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /api/interview-sessions/{sessionId}/scoring/retry:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    post:
      tags: [Interview]
      operationId: retryInterviewScoring
      summary: Retry failed scoring
      responses:
        '202':
          description: Scoring redispatched.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewReport'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /api/admin/users:
    get:
      tags: [Admin]
      operationId: listAdminUsers
      summary: Search and filter user accounts
      parameters:
        - name: keyword
          in: query
          required: false
          schema: {type: string}
        - name: role
          in: query
          required: false
          schema: {$ref: '#/components/schemas/UserRole'}
        - name: enabled
          in: query
          required: false
          schema: {type: boolean}
        - name: page
          in: query
          required: false
          schema: {type: integer, minimum: 0, default: 0}
        - name: size
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: Matching accounts ordered newest first.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminUserPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/admin/users/{userId}:
    parameters:
      - {$ref: '#/components/parameters/UserId'}
    get:
      tags: [Admin]
      operationId: getAdminUser
      summary: Get account details and activity counts
      responses:
        '200':
          description: Account details.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminUserDetail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/admin/users/{userId}/status:
    parameters:
      - {$ref: '#/components/parameters/UserId'}
    patch:
      tags: [Admin]
      operationId: updateAdminUserStatus
      summary: Enable or disable a USER account
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateUserStatusRequest'}
      responses:
        '200':
          description: Updated account. Disabling also revokes all refresh tokens.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminUserDetail'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/admin/interview-sessions:
    get:
      tags: [Admin]
      operationId: listAdminInterviewSessions
      summary: Search and filter all interview sessions
      parameters:
        - name: keyword
          in: query
          required: false
          description: Case-insensitive owner name or email search.
          schema: {type: string}
        - name: status
          in: query
          required: false
          schema: {$ref: '#/components/schemas/InterviewSessionStatusEnum'}
        - name: mode
          in: query
          required: false
          schema: {$ref: '#/components/schemas/InterviewSessionMode'}
        - name: from
          in: query
          required: false
          schema: {type: string, format: date-time}
        - name: to
          in: query
          required: false
          schema: {type: string, format: date-time}
        - name: page
          in: query
          required: false
          schema: {type: integer, minimum: 0, default: 0}
        - name: size
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: Matching sessions ordered newest first.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminSessionPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/admin/interview-sessions/{sessionId}:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    get:
      tags: [Admin]
      operationId: getAdminInterviewSession
      summary: Get interview-session operational details
      responses:
        '200':
          description: Session metadata, processing errors, and transition history.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminSessionDetail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/admin/interview-sessions/{sessionId}/preparation/retry:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    post:
      tags: [Admin]
      operationId: retryAdminInterviewPreparation
      summary: Retry failed interview preparation as an administrator
      responses:
        '202':
          description: Preparation redispatched.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminSessionDetail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/admin/interview-sessions/{sessionId}/scoring/retry:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    post:
      tags: [Admin]
      operationId: retryAdminInterviewScoring
      summary: Retry failed interview scoring as an administrator
      responses:
        '202':
          description: Scoring redispatched.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminSessionDetail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  parameters:
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
    UserId:
      name: userId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
  responses:
    BadRequest:
      description: Invalid input.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Unauthorized:
      description: Missing or invalid authentication.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Forbidden:
      description: Authenticated but not allowed.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    NotFound:
      description: Resource not found or inaccessible.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Conflict:
      description: State, idempotency, version, or limit conflict.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
...
  schemas:
    ApiError:
      type: object
      required: [timestamp, status, code, message, path]
      properties:
        timestamp: {type: string, format: date-time}
        status: {type: integer}
        code: {type: string}
        message: {type: string}
        path: {type: string}
        fieldErrors:
          type: object
          nullable: true
          additionalProperties: {type: string}
    UserRole: {type: string, enum: [ADMIN, USER]}
...
    AdminUserSummary:
      type: object
      required: [id, fullName, email, role, enabled, createdAt, updatedAt]
      properties:
        id: {type: integer, format: int64}
        fullName: {type: string}
        email: {type: string, format: email}
        role: {$ref: '#/components/schemas/UserRole'}
        enabled: {type: boolean}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    AdminUserActivity:
      type: object
      required: [activeCvCount, activeJobDescriptionCount, interviewSessionCount, completedInterviewSessionCount]
      properties:
        activeCvCount: {type: integer, format: int64, minimum: 0}
        activeJobDescriptionCount: {type: integer, format: int64, minimum: 0}
        interviewSessionCount: {type: integer, format: int64, minimum: 0}
        completedInterviewSessionCount: {type: integer, format: int64, minimum: 0}
    AdminUserDetail:
      type: object
      required: [id, fullName, email, role, enabled, createdAt, updatedAt, activity]
      properties:
        id: {type: integer, format: int64}
        fullName: {type: string}
        email: {type: string, format: email}
        avatarUrl: {type: string, format: uri, nullable: true}
        role: {$ref: '#/components/schemas/UserRole'}
        enabled: {type: boolean}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        activity: {$ref: '#/components/schemas/AdminUserActivity'}
    UpdateUserStatusRequest:
      type: object
      required: [enabled]
      properties:
        enabled: {type: boolean}
    AdminUserPage:
      type: object
      required: [items, page, size, totalElements]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/AdminUserSummary'}
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1}
        totalElements: {type: integer, format: int64, minimum: 0}
    AdminUserReference:
      type: object
      required: [id, fullName, email]
      properties:
        id: {type: integer, format: int64}
        fullName: {type: string}
        email: {type: string, format: email}
    AdminSessionSummary:
      type: object
      required: [id, user, status, templateTitle, profileName, mode, languageCode, durationMinutes, createdAt, updatedAt]
      properties:
        id: {type: integer, format: int64}
        user: {$ref: '#/components/schemas/AdminUserReference'}
        status: {$ref: '#/components/schemas/InterviewSessionStatusEnum'}
        templateTitle: {type: string}
        profileName: {type: string}
        mode: {$ref: '#/components/schemas/InterviewSessionMode'}
        languageCode: {type: string}
        durationMinutes: {type: integer, minimum: 1}
        preparationErrorCode: {type: string, nullable: true}
        scoringErrorCode: {type: string, nullable: true}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    AdminSessionPage:
      type: object
      required: [items, page, size, totalElements]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/AdminSessionSummary'}
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1}
        totalElements: {type: integer, format: int64, minimum: 0}
    InterviewTransitionActor:
      type: string
      enum: [USER, ADMIN, SYSTEM, SCHEDULER]
    AdminSessionTransition:
      type: object
      required: [id, toStatus, actor, occurredAt]
      properties:
        id: {type: integer, format: int64}
        fromStatus:
          allOf: [{$ref: '#/components/schemas/InterviewSessionStatusEnum'}]
          nullable: true
        toStatus: {$ref: '#/components/schemas/InterviewSessionStatusEnum'}
        reason: {type: string, nullable: true}
        actor: {$ref: '#/components/schemas/InterviewTransitionActor'}
        occurredAt: {type: string, format: date-time}
    AdminSessionDetail:
      type: object
      required: [id, user, status, templateTitle, profileName, languageCode, durationMinutes, interviewerStyle, mode, currentTurnIndex, createdAt, updatedAt, transitions]
      properties:
        id: {type: integer, format: int64}
        user: {$ref: '#/components/schemas/AdminUserReference'}
        status: {$ref: '#/components/schemas/InterviewSessionStatusEnum'}
        templateTitle: {type: string}
        profileName: {type: string}
        languageCode: {type: string}
        durationMinutes: {type: integer, minimum: 1}
        interviewerStyle: {$ref: '#/components/schemas/InterviewerStyle'}
        mode: {$ref: '#/components/schemas/InterviewSessionMode'}
        realtimeProvider: {type: string, nullable: true}
        realtimeVoiceName: {type: string, nullable: true}
        currentTurnIndex: {type: integer, minimum: 0}
        preparationErrorCode: {type: string, nullable: true}
        preparationErrorMessage: {type: string, nullable: true}
        planSchemaVersion: {type: string, nullable: true}
        planModelName: {type: string, nullable: true}
        planPromptVersion: {type: string, nullable: true}
        scoringErrorCode: {type: string, nullable: true}
        scoringErrorMessage: {type: string, nullable: true}
        preparationStartedAt: {type: string, format: date-time, nullable: true}
        preparedAt: {type: string, format: date-time, nullable: true}
        startedAt: {type: string, format: date-time, nullable: true}
        deadlineAt: {type: string, format: date-time, nullable: true}
        lastActivityAt: {type: string, format: date-time, nullable: true}
        endReason: {$ref: '#/components/schemas/InterviewEndReason'}
        endedAt: {type: string, format: date-time, nullable: true}
        scoringStartedAt: {type: string, format: date-time, nullable: true}
        completedAt: {type: string, format: date-time, nullable: true}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        transitions:
          type: array
          items: {$ref: '#/components/schemas/AdminSessionTransition'}
...
    InterviewerStyle: {type: string, enum: [FRIENDLY, PROFESSIONAL, CHALLENGING]}
    InterviewSessionMode: {type: string, enum: [TURN_BASED, VOICE_REALTIME]}
...
    InterviewSessionStatusEnum:
      type: string
      enum: [PREPARING, READY, PREPARATION_FAILED, IN_PROGRESS, SCORING, COMPLETED, SCORING_FAILED, CANCELLED, EXPIRED]
    InterviewEndReason:
      type: string
      enum: [AI_COMPLETED, TIME_EXPIRED, CANDIDATE_FINISHED, SYSTEM_TERMINATED]
      nullable: true
...
    InterviewSessionStatus:
      type: object
      required: [id, status, templateTitle, profileName, languageCode, durationMinutes, interviewerStyle, mode]
      properties:
        id: {type: integer, format: int64}
        status: {$ref: '#/components/schemas/InterviewSessionStatusEnum'}
        templateTitle: {type: string}
        profileName: {type: string}
        languageCode: {type: string}
        durationMinutes: {type: integer}
        interviewerStyle: {$ref: '#/components/schemas/InterviewerStyle'}
        mode: {$ref: '#/components/schemas/InterviewSessionMode'}
        realtimeProvider: {type: string, nullable: true}
        realtimeVoiceName: {type: string, nullable: true}
        preparationErrorCode: {type: string, nullable: true}
        preparationErrorMessage: {type: string, nullable: true}
        scoringErrorCode: {type: string, nullable: true}
        scoringErrorMessage: {type: string, nullable: true}
        preparedAt: {type: string, format: date-time, nullable: true}
        endReason: {$ref: '#/components/schemas/InterviewEndReason'}
        endedAt: {type: string, format: date-time, nullable: true}
        completedAt: {type: string, format: date-time, nullable: true}
...
    InterviewReport:
      type: object
      required: [sessionId, status, report]
      properties:
        sessionId: {type: integer, format: int64}
        status: {$ref: '#/components/schemas/InterviewSessionStatusEnum'}
        scoringErrorCode: {type: string, nullable: true}
        scoringErrorMessage: {type: string, nullable: true}
        report:
          allOf: [{$ref: '#/components/schemas/InterviewReportContent'}]
          nullable: true
        completedAt: {type: string, format: date-time, nullable: true}
    InterviewReportContent:
      type: object
      required: [summary, scores, focusAreas, recommendations]
      properties:
        score: {type: number, nullable: true}
        summary: {type: string, maxLength: 400}
        scores: {$ref: '#/components/schemas/InterviewScoreBreakdown'}
        focusAreas:
          type: array
          items: {$ref: '#/components/schemas/InterviewFocusAreaScore'}
        recommendations:
          type: array
          maxItems: 3
          items: {type: string, maxLength: 200}
    InterviewScoreBreakdown:
      type: object
      required: [technical, communication]
      properties:
        technical: {$ref: '#/components/schemas/InterviewScoreFeedback'}
        communication: {$ref: '#/components/schemas/InterviewScoreFeedback'}
    InterviewScoreFeedback:
      type: object
      required: [feedback]
      properties:
        score: {type: number, nullable: true}
        feedback: {type: string, maxLength: 300}
    InterviewFocusAreaScore:
      type: object
      required: [name]
      properties:
        name: {type: string}
        score: {type: number, nullable: true}
security:
  - bearerAuth: []